% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniRead.R
\name{aniRead}
\alias{aniRead}
\title{Parse and Read-In ani Output files}
\usage{
aniRead(output.filepath = paste0(getwd(), "/output/output.out.nsub"))
}
\arguments{
\item{output.filepath}{Output file path to read in}
}
\value{
output.output
}
\description{
Parses the animation file for data usage (either per sub-basin, hru, or others)
}
