% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniReadSubbasin.R
\name{aniReadSubbasin}
\alias{aniReadSubbasin}
\title{aniRead for Subbasins}
\usage{
aniReadSubbasin(output.filepath = paste0(getwd(),
  "/output/output.out.nsub"))
}
\arguments{
\item{output.filepath}{Output file path to read in}
}
\value{
output.output
}
\description{
This is a modification of the aniRead function to work with nsub instead of nHRUs.
}
