% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRead.R
\name{ControlRead}
\alias{ControlRead}
\title{Read in Control File for PRMS}
\usage{
ControlRead(control.filepath)
}
\arguments{
\item{control.filepath}{Control file path to read in}
}
\value{
output
}
\description{
Read in Control File for PRMS
}
