% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlWrite.R
\name{ControlWrite}
\alias{ControlWrite}
\title{Generates control file for PRMS}
\usage{
ControlWrite(control.values, control.filepath, csv = FALSE)
}
\arguments{
\item{control.values}{Control Information File}

\item{control.filepath}{File Path and File Name for Control File}

\item{csv}{Determine if generate as CSV}
}
\description{
Generates control file for PRMS
}
