% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFileWrite.R
\name{DataFileWrite}
\alias{DataFileWrite}
\title{Writes Data File}
\usage{
DataFileWrite(datainput.data, datainput.writepath)
}
\arguments{
\item{datainput.data}{List of Variables and Data}

\item{datainput.writepath}{Input data file writing path}
}
\description{
Writes Data File
}
