% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamWrite.R
\name{ParamWrite}
\alias{ParamWrite}
\title{Generates parameter file for PRMS}
\usage{
ParamWrite(prms.param.values, param.path)
}
\arguments{
\item{prms.param.values}{PRMS Parameter Files values}

\item{param.path}{File Path and File Name for Parameter File}
}
\description{
Generates parameter file for PRMS
}
