% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingControl.R
\name{PreprocessingControl}
\alias{PreprocessingControl}
\title{Preparing and updating control values}
\usage{
PreprocessingControl(control.input)
}
\arguments{
\item{control.input}{Control Input List}
}
\value{
control.input
}
\description{
Preparing and updating control values
}
