% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prmsR.R
\docType{package}
\name{prmsR-package}
\alias{prmsR}
\alias{prmsR-package}
\title{prmsR Package}
\description{
Maintained by Hydrosystems Research Group at University of Massachusetts Amherst.
}
\details{
prmsR is a small internal library that operates as a wrapper, parser, and file generator
for the PRMS-IV model by USGS.
}
\author{
\strong{Maintainer}: Don Park \email{donpark@umass.edu}

}
