% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prmsRun.R
\name{prmsRun}
\alias{prmsRun}
\title{PRMS Run Wrapper}
\usage{
prmsRun(binary.filepath, control.filepath)
}
\arguments{
\item{binary.filepath}{Filepath to the PRMS Binary (Relative or Absolute)}

\item{control.filepath}{Filepath to the control file (Relative or Absolute)}
}
\description{
PRMS Run Wrapper
}
