% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatVarRead.R
\name{StatVarRead}
\alias{StatVarRead}
\title{Parse and Read-In Statvar Output file}
\usage{
StatVarRead(output.filepath = paste0(getwd(), "/output/output.statvar"))
}
\arguments{
\item{output.filepath}{Output file path to read in}
}
\value{
output.output
}
\description{
Parse and Read-In Statvar Output file
}
